- rm utf8 comment from modules - update module docstrings - rename modules with API classes for better distinction
