+++ src/UniProtMapper/__init__.py
@@ -1,7 +1,8 @@
-# -*- coding: utf-8 -*-
 """A Python wrapper for the UniProt id-mapping RESTful API: https://www.uniprot.org/id-mapping"""
+
 try:
-    from .api import ProtMapper  # noqa: F401
+    from .idmapping_api import ProtMapper  # noqa: F401
+    from .uniprotkb_api import ProtKB  # noqa: F401
 except ImportError:  # when installing the package with setuptools_scm
     pass
