+++ src/UniProtMapper/uniprotkb_fields.py
